+++ src/connectedk8s/azext_connectedk8s/_constants.py
@@ -115,17 +115,18 @@ Default_Namespace_Does_Not_Exist_Fault_Type = "The default namespace defined in 
 ClusterConnect_Not_Present_Fault_Type = "cluster-connect-feature-unavailable"
 KAP_1P_Server_App_Scope = "6256c85f-0aad-4d50-b960-e6e9b21efe35/.default"
 KAP_1P_Server_AppId = "6256c85f-0aad-4d50-b960-e6e9b21efe35"
-Get_PublicKey_Info_Fault_Type = 'Error while fetching the PoP publickey information from CSP'
+Get_PublicKey_Info_Fault_Type = 'Error while fetching the PoP publickey information from client proxy'
 PoP_Public_Key_Expried_Fault_Type = 'The PoP public key used to generate the at has expired'
+Post_AT_To_ClientProxy_Failed_Fault_Type = 'Failed to post access token to client proxy'
 AZ_CLI_ADAL_TO_MSAL_MIGRATE_VERSION = '2.30.0'
-CLIENT_PROXY_VERSION = '3'
+CLIENT_PROXY_VERSION = '1.3.018802'
 API_SERVER_PORT = 47011
 CLIENT_PROXY_PORT = 47010
 CLIENTPROXY_CLIENT_ID = '04b07795-8ddb-461a-bbee-02f9e1bf7b46'
 API_CALL_RETRIES = 12
 DEFAULT_REQUEST_TIMEOUT = 10  # seconds
-RELEASE_DATE_WINDOWS = 'release16-12-21'
-RELEASE_DATE_LINUX = 'release16-12-21'
+RELEASE_DATE_WINDOWS = 'release03-04-22'
+RELEASE_DATE_LINUX = 'release03-04-22'
 CSP_REFRESH_TIME = 300
 # URL constants
 CSP_Storage_Url = "https://k8sconnectcsp.azureedge.net"
